pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "Evrat",email = "Evratal@yandex.ru"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "dispatcher (>=1.0,<2.0)",
    "pytest (>=8.3.5,<9.0.0)"
]

[tool.poetry]
packages = [{include = "config", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
